+++ src/test/java/org/apache/sysds/test/component/compress/CompressedTestBase.java
@@ -336,7 +336,7 @@ public abstract class CompressedTestBase extends TestBase {
 	 * @param valueRange The value range used as input
 	 */
 	private void setLossyTolerance(ValueRange valueRange) {
-		lossyTolerance = (double) (Math.max(TestConstants.getMaxRangeValue(valueRange),
+		lossyTolerance = (Math.max(TestConstants.getMaxRangeValue(valueRange),
 			Math.abs(TestConstants.getMinRangeValue(valueRange)))) * (1.0 / 127.0) / 2.0;
 	}
